rename: chatProvider -> NotificationProvider
